requirements.txt
aiogram>=3.10,<4.0
google-genai>=1.0
aiosqlite>=0.20
apscheduler>=3.10
notion-client>=2.2
python-dotenv>=1.0
Pillow>=10.0
pydantic>=2.7,<2.14
httpx>=0.27
